requirements.txt
# WhoseOnFirst - Python Dependencies
# Production requirements

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.1

# Database
sqlalchemy==2.0.31
alembic==1.13.1  # Database migrations

# Scheduling
apscheduler==3.10.4
pytz==2024.1

# SMS Provider
twilio==9.2.3

# Data Validation
pydantic==2.8.2
pydantic-settings==2.2.1

# Utilities
python-dotenv==1.0.1
httpx==0.25.2  # Async HTTP client

# Security
argon2-cffi==23.1.0  # Argon2id password hashing (OWASP 2025 recommended)

# Logging
python-json-logger==2.0.7
